third_party/blink/web_tests/fast/sub-pixel/boundingclientrect-subpixel-margin-expected.txt

PASS testRect.left - parentRect.left is 1.5
PASS parentRect.right - testRect.right is 1.5
PASS successfullyParsed is true

TEST COMPLETE
